Updated DB handler to allow JSON to be recieved to both edit and create objects. Also sends back JSON when in JSON mode and create/update was handled.
Additionally, the DB handler now handles authorization on the property-level.
This was previously done by XSLT, but now that we allow JSON, this simply can't be
done all the time. Unforunately XSLT will still be the source for authentication
on the property level for sending back params when working in XML mode, but
eventually XML mode should be depricated.