Updated DB handler to override the __call__ method to trap SDBResponseErrors and toss them as BadRequests.
Updated the "Report Exception" function to simply use str(exception)
instead of looking for the "message" explicitly. This is more
useful when dealing with any exception generically

Updated the base HTTPException class to also include the "Description"
in the string version. This helps so we can see the actual
query after the first change in this changeset